Move compatibility stuff into utils/compat
