+++ astropy/io/vo/util.py
@@ -37,14 +37,14 @@ def convert_to_writable_filelike(fd):
     if isinstance(fd, basestring):
         if IS_PY3K:
             if fd.endswith('.gz'):
-                from ...utils import gzip
+                from ...utils.compat import gzip
                 fd = gzip.GzipFile(fd, 'wb')
                 fd = io.TextIOWrapper(fd, encoding='utf8')
             else:
                 fd = io.open(fd, 'w', encoding='utf8')
         else:
             if fd.endswith('.gz'):
-                from ...utils import gzip
+                from ...utils.compat import gzip
                 fd = gzip.GzipFile(fd, 'wb')
             else:
                 fd = open(fd, 'wb')
@@ -95,7 +95,7 @@ def convert_to_fd_or_read_function(fd):
         return fd
     elif isinstance(fd, basestring):
         if fd.endswith('.gz'):
-            from ...utils import gzip
+            from ...utils.compat import gzip
             fd = gzip.GzipFile(fd, 'rb')
             return fd.read
         else:
