+++ astropy/utils/compat/gzip.py
